// webpack---_wix_analytics_ng_bm/constants/enums.ts
export enum DashboardModeBI {
  regular = 'regular',
  advanced = 'advanced',
}
// ...
export enum AddRemoveStatActionBI {
  add = 'add',
  remove = 'remove',
}
// ...
export enum ExperimentNames {
  analyticsNGDashboardWidgetNext = 'specs.analytics-ng.AnalyticsNGDashboardWidgetNext',
  useNewDashboard2020 = 'specs.wos1.useNewDashboard2020',
  customDashboardDownload = 'specs.analytics-ng.CustomDashboardDownload',
  insights = 'specs.analytics-ng.Insights',
  pngDownload = 'specs.analytics-ng.PNGDownload',
  pngDownloadForTables = 'specs.analytics-ng.PNGDownloadForTables',
  cacheOverride = 'specs.analytics-ng.CacheOverride',
  download = 'specs.analytics-ng.Download',
  benchmarks = 'specs.analytics-ng.Benchmarks',
  siteSpeed = 'specs.analytics-ng.SiteSpeed',
  analyticsNGInsightsWidget = 'specs.analytics-ng.AnalyticsNGInsightsWidget',
  subscriptions = 'specs.analytics-ng.Subscriptions',
  subscriptionsPhase2 = 'specs.analytics-ng.SubscriptionsPhase2',
  benchmarkActions = 'specs.analytics-ng.BenchmarkActions',
  alerts = 'specs.analytics-ng.Alerts',
  newReportSubscriptionModal = 'specs.analytics-ng.NewReportSubscriptionModal',
  useWixSubdomain = 'specs.analytics-ng.UseWixSubdomain',
  nativeOverviewSales = 'specs.analytics-ng.NativeOverviewSales',
  insightsV3 = 'specs.analytics-ng.InsightsV3',
  forecast = 'specs.analytics-ng.Forecast',
  radarChartWSR = 'specs.analytics-ng.RadarChartWSR',
  disableNinetyDaysFilterWidget = 'specs.analytics-ng.disableNinetyDaysFilterWidget',
  enableConversionReports = 'specs.analytics-ng.EnableConversionReports',
  forecastEntry = 'specs.analytics-ng.ForecastEntry',
}
// ...
export enum DashboardWidgetType {
  horizontal = 'horizontal',
  vertical = 'vertical',
}